pyproject.toml
[project]
name = "roadeye"
version = "1.0.0"
description = "Wildlife roadkill detection and mapping platform"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [
    {name = "RoadEye Team"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: GIS",
]

dependencies = [
    # Core
    "pydantic>=2.0,<3.0",
    "pydantic-settings>=2.0,<3.0",
    "pyyaml>=6.0",

    # Database
    "sqlalchemy>=2.0,<3.0",
    "alembic>=1.13,<2.0",

    # API
    "fastapi>=0.109,<1.0",
    "uvicorn[standard]>=0.27,<1.0",
    "python-multipart>=0.0.6",
    "httpx>=0.26",

    # Detection
    "ultralytics>=8.1",
    "opencv-python-headless>=4.9",
    "numpy>=1.24,<3.0",

    # Analytics
    "pandas>=2.0,<3.0",
    "scikit-learn>=1.4,<2.0",

    # Visualisation
    "folium>=0.15",
    "jinja2>=3.1",

    # CLI
    "click>=8.1,<9.0",
    "rich>=13.0",

    # Logging
    "structlog>=24.1",
]

[project.optional-dependencies]
edge = [
    "opencv-python>=4.9",
    "requests>=2.31",
]
training = [
    "roboflow>=1.1",
    "albumentations>=1.3",
]
geo = [
    "geopandas>=0.14",
]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "factory-boy>=3.3",
    "faker>=22.0",
    "mypy>=1.8",
    "ruff>=0.1",
    "pre-commit>=3.6",
]

[project.scripts]
roadeye = "roadkill.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/roadkill"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = [
    "--strict-markers",
    "-ra",
]

[tool.ruff]
target-version = "py39"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "C4", "SIM"]
ignore = [
    "E501",   # Line too long (handled by formatter)
    "B008",   # Do not perform function call in defaults - FastAPI Depends pattern
    "UP006",  # Use `dict` instead of `Dict` - keep for 3.9 compat
    "UP007",  # Use `X | Y` instead of `Union` - keep for 3.9 compat
    "UP035",  # Deprecated typing imports - keep for 3.9 compat
    "UP037",  # Remove quotes from annotations - keep for 3.9 compat
    "UP045",  # Use `X | None` instead of `Optional` - keep for 3.9 compat
]

[tool.ruff.lint.isort]
known-first-party = ["roadkill"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_ignores = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = [
    "ultralytics.*",
    "cv2.*",
    "folium.*",
    "sklearn.*",
    "pandas.*",
    "geopandas.*",
]
ignore_missing_imports = true
